requirements.txt
dotenv==0.9.9
PyGithub==2.7.0
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2